% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{add_all_packages}
\alias{add_all_packages}
\title{Update all (active) CRAN packages}
\usage{
add_all_packages()
}
\value{
A list of results. For successful updates \code{TRUE}.
  For unsuccessful ones, a \code{try-error} object.
}
\description{
Update all (active) CRAN packages
}

