% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{add_gh_remote}
\alias{add_gh_remote}
\title{Add a CRAN at GitHub remote to a local git tree}
\usage{
add_gh_remote(package)
}
\arguments{
\item{package}{Package name.}
}
\description{
It is assumed that the git tree of the package is
in the working directory.
}
\details{
If there is already a remote called \code{origin}, then it does nothing.
}
\keyword{internal}
