% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-package.R
\name{add_missing_version}
\alias{add_missing_version}
\title{Add a single missing version of a package to the GitHub mirror}
\usage{
add_missing_version(package, version, date)
}
\arguments{
\item{package}{Name of the package to update.}

\item{version}{Character scalar, the version to add.}

\item{date}{The exact date and time the version was built for CRAN.}
}
\value{
The package's metadata, i.e. the contents of the \code{DESCRIPTION}
file, in a \code{desciption} object, see the \code{description} package.
}
\description{
It assumes that the package's repository exists on GitHub,
and that the working directory contains a clone of of it.
}
\details{
It then downloads the specified version from CRAN and adds it
to the local clone. Finally, it pushes the repo to GitHub.
}
\keyword{internal}
