% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_missing_versions}
\alias{add_missing_versions}
\title{Add some (or all) versions of a package to the GitHub mirror}
\usage{
add_missing_versions(package, versions, new_package, timeline)
}
\arguments{
\item{package}{Name of the package to update.}

\item{versions}{Character vector, package versions to add.}

\item{new_package}{Logical scalar, whether the package is new. If
the package is new, then its repo does not exists (yet) on GitHub.}

\item{timeline}{The full timeline of the package, from crandb.}
}
\description{
Add some (or all) versions of a package to the GitHub mirror
}
\keyword{internal}

