% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_package}
\alias{add_package}
\title{Add or update a package in CRAN @ GitHub}
\usage{
add_package(package)
}
\arguments{
\item{package}{Name of the package}
}
\value{
Invisible `TRUE` if the package was successfully updated,
  `FALSE` otherwise.
}
\description{
Add or update a package in CRAN @ GitHub
}

