% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_package}
\alias{check_package}
\title{Check if the CRAN@GH mirror of a package is up to date}
\usage{
check_package(package)
}
\arguments{
\item{package}{Name of the package.}
}
\value{
Character vector.
}
\description{
It returns a character vector of the missing version numbers. If
a package is up to date, then an empty character vector is returned.
}
