% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{clean_description}
\alias{clean_description}
\title{Clean a string, so that it can be used as a repo description}
\usage{
clean_description(description)
}
\arguments{
\item{description}{Character scalar.}
}
\value{
Character scalr. Cleaned up description.
}
\description{
Currently it just removes newline characters.
}
\keyword{internal}
