% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{clone_git_repo}
\alias{clone_git_repo}
\title{Clone a repository from the GitHub CRAN mirror}
\usage{
clone_git_repo(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
Output of the command line git call.
}
\description{
Clone a repository from the GitHub CRAN mirror
}
\keyword{internal}
