% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\docType{data}
\name{cran_mirror}
\alias{cran_mirror}
\title{The CRAN mirror to use for package downloads}
\format{An object of class \code{character} of length 1.}
\usage{
cran_mirror
}
\description{
The CRAN mirror to use for package downloads
}
\keyword{internal}
