% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb.R
\name{cran_url}
\alias{cran_url}
\title{Get the URL of a CRAN package}
\usage{
cran_url(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
Character scalar, the URL.
}
\description{
Get the URL of a CRAN package
}
\keyword{internal}
