% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{create_gh_repo}
\alias{create_gh_repo}
\title{Create a repository for a package at GitHub}
\usage{
create_gh_repo(package, description = make_description(package))
}
\arguments{
\item{package}{Package name.}

\item{description}{Repository description.}
}
\description{
Create a repository for a package at GitHub
}
\keyword{internal}
