% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{create_git_repo}
\alias{create_git_repo}
\title{Create an empty local git tree}
\usage{
create_git_repo(package)
}
\arguments{
\item{package}{Package name, directory to create a git tree in.}
}
\value{
The return value of the call to git.
}
\description{
Create an empty local git tree
}
\keyword{internal}
