% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dont_break}
\alias{dont_break}
\title{Replace are whitespace with non-breakable space in a string}
\usage{
dont_break(...)
}
\arguments{
\item{...}{Input, it will be concatenated using \code{\link[base:paste]{base::paste0()}}.}
}
\value{
Character.
}
\description{
Replace are whitespace with non-breakable space in a string
}
\keyword{internal}
