% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{get_all_cran_packages}
\alias{get_all_cran_packages}
\title{Get the names and version numbers of all CRAN packages}
\usage{
get_all_cran_packages()
}
\value{
Character vector of all package names.
}
\description{
Including archived ones.
}
\keyword{internal}
