% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_clone_url}
\alias{get_clone_url}
\title{The clone URL of a package at GitHub}
\usage{
get_clone_url(package)
}
\arguments{
\item{package}{Name of the package.}
}
\value{
The url.
}
\description{
It is always an HTTPS URL currently.
}
\keyword{internal}
