% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb.R
\name{get_crandb_timeline}
\alias{get_crandb_timeline}
\title{Get the timeline of a package from crandb}
\usage{
get_crandb_timeline(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
The timeline, in a named list, where names are
  package versions, and entries are dates. If the package is not
  in crandb, \code{NULL} is returned. A special entry named
  \code{archived} is included for archived packages.
}
\description{
Get the timeline of a package from crandb
}
\keyword{internal}
