% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_gh_owner}
\alias{get_gh_owner}
\title{The GitHub repository or organization of the mirror}
\usage{
get_gh_owner(package)
}
\value{
Character scalar.
}
\description{
The GitHub repository or organization of the mirror
}
\keyword{internal}

