% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{get_github_versions}
\alias{get_github_versions}
\title{Get all versions of a package in the GitHub mirror}
\usage{
get_github_versions(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
Character vector of version numbers.
}
\description{
Get all versions of a package in the GitHub mirror
}
\keyword{internal}
