% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get_mirror_status}
\alias{get_mirror_status}
\title{Check if all packages are up to date on the CRAN@GH mirror}
\usage{
get_mirror_status()
}
\value{
A data frame with columns \code{package}, \code{version_cran},
\code{version_github}, \code{status}. If a package is missing from CRAN/GitHub,
then its version is \code{NA}. Possible values of \code{status} are
\itemize{
\item \code{"ok"}: mirror is current,
\item \code{"removed"}: still on the mirror, but it was removed from CRAN,
\item \code{"missing"}: missing from the mirror entirely,
\item \code{"outdated"}: outdated on the mirror.
}
}
\description{
Check if all packages are up to date on the CRAN@GH mirror
}
