% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_package_tarball}
\alias{get_package_tarball}
\title{Download CRAN source package into the current directory}
\usage{
get_package_tarball(package, version)
}
\arguments{
\item{package}{Package name.}

\item{version}{Package version.}
}
\value{
Name of the downloaded file
}
\description{
It puts the downloaded tarball into the current working directory.
}
\keyword{internal}
