% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\title{Run a git command using system git}
\usage{
git(..., env = character(), timeout = 60 * 60)
}
\arguments{
\item{...}{Arguments to pass to git.}

\item{env}{Named character vector, environment variables
to set for git.}

\item{timeout}{Timeout, in seconds, defaults to one hour.}
}
\value{
Return value from \code{\link[processx:run]{processx::run()}}.
}
\description{
System git must be properly installed and it must be in the \code{PATH}.
}
\details{
It throws an error on error.
}
\keyword{internal}
