% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{make_description}
\alias{make_description}
\title{Create a GitHub description string for a package}
\usage{
make_description(pkg)
}
\arguments{
\item{pkg}{Package name, or a `description` object.}
}
\value{
Character scalar, an appropriate GitHub repository
  description.
}
\description{
On GitHub the description cannot contain newline characters,
and HTML tags are not allowed, either. But it is possible to include
non-breaking space characters, and this results for a nice
multi-line description for most packages.
}
\keyword{internal}
