% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb.R
\name{metacran_url}
\alias{metacran_url}
\title{Get the Metacran (\url{http://www.r-pkg.org} URL of a package}
\usage{
metacran_url(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
The Metacran URL.
}
\description{
Get the Metacran (\url{http://www.r-pkg.org} URL of a package
}
\keyword{internal}

