% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_empty}
\alias{na_to_empty}
\title{Empty sting from \code{NA}}
\usage{
na_to_empty(x)
}
\arguments{
\item{x}{Input.}
}
\value{
\code{x} or and empty string.
}
\description{
If \code{x} is a single \code{NA} (any mode),
and empty string is returned. Otherwise \code{x}
is returned.
}
\keyword{internal}
