% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullna_or}
\alias{nullna_or}
\title{Empty string if input is \code{NULL}.}
\usage{
nullna_or(x, expr)
}
\arguments{
\item{x}{Condition.}

\item{expr}{Expression.}
}
\value{
Expression or an empty string.
}
\description{
If \code{x} is \code{NULL}, then an empty string is returned,
otherwise \code{expr}.
}
\keyword{internal}

