% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{package_urls}
\alias{package_urls}
\title{Possible download URLs for CRAN packages}
\usage{
package_urls(package, version)
}
\arguments{
\item{package}{Package name.}

\item{version}{Package version.}
}
\value{
Character vector of URLs.
}
\description{
Possible download URLs for CRAN packages
}
\keyword{internal}

