% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{push_to_github}
\alias{push_to_github}
\title{Push the package to GitHub}
\usage{
push_to_github(package, forced_push = FALSE)
}
\arguments{
\item{package}{Package name.}

\item{forced_push}{Whether to make a forced push.}
}
\value{
Output of the command line git call.
}
\description{
Push the package to GitHub
}
\keyword{internal}
