% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh_status_yml}
\alias{refresh_status_yml}
\title{Refresh cranatgh-status.yml}
\usage{
refresh_status_yml()
}
\description{
Do a full refresh of \code{cranatgh-status.yml} by scraping all github
repositories from the CRAN organization.
}
