% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{remove_gh_repo}
\alias{remove_gh_repo}
\title{Remove a GitHub remote}
\usage{
remove_gh_repo(package)
}
\arguments{
\item{package}{Package name.}
}
\description{
This is only used for the unit tests. Otherwise we don't remove
repositories from GitHub, even if a package is archived.
}
\keyword{internal}

