% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{setup_repo_features}
\alias{setup_repo_features}
\title{Set up the features of the GitHub repos}
\usage{
setup_repo_features(pkgs = NULL)
}
\arguments{
\item{pkgs}{Packages to update. Defaults to all packages.}
}
\description{
Including:
\itemize{
\item set repo description,
\item set repo link,
\item disable issues,
\item disable wiki,
\item disable project.
}
}
