% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{try_download}
\alias{try_download}
\title{Try a package download from CRAN}
\usage{
try_download(url, dest_file)
}
\arguments{
\item{url}{Download url}

\item{dest_file}{Destination file}
}
\value{
Logical scalar, whether the download was successfull.
}
\description{
It tries to write the file directly on disk.
Unsuccessful downloads result an error.
}
\keyword{internal}
