% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onetime.R
\name{update_all_descriptions}
\alias{update_all_descriptions}
\title{Update the repository description of some packages}
\usage{
update_all_descriptions(pkgs = get_all_cran_packages(), sleep = 1)
}
\arguments{
\item{pkgs}{Packages to update. By default all active CRAN
packages are updated.}

\item{sleep}{Numeric scalar, the amount of time to wait between
GitHub API calls, in seconds. Defaults to one second.}
}
\value{
A list of results. For successful updates the
  response of the GitHub API call. For unsuccessful ones,
  a \code{try-error} object.
}
\description{
Update the repository description of some packages
}

