% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{update_all_packages}
\alias{update_all_packages}
\title{Update all missing and outdated packages}
\usage{
update_all_packages(max_failures = 10, update_cache_after = 100)
}
\arguments{
\item{max_failures}{Maximum number of failures. Many failures will
cause the updates to stop.}

\item{update_cache_after}{Update cache after this many iterations.}
}
\value{
Data frame, similar to the return value of \code{\link[=get_mirror_status]{get_mirror_status()}},
with the extra \code{result} column, which is \code{"success"} for successful,
\code{"failure"} for unsuccessful updated, and \code{NA} otherwise.
}
\description{
Update all missing and outdated packages
}
