% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{update_description}
\alias{update_description}
\title{Update description on GitHub}
\usage{
update_description(package, description)
}
\arguments{
\item{package}{Name of the package}

\item{description}{\code{Title} field in the \code{DESCRIPTION} file.}
}
\description{
Updates the repostory description and homepage.
}

