% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_tempdir}
\alias{with_tempdir}
\title{Run an expression in a temporary directory}
\usage{
with_tempdir(expr, tmpdir = tempfile(), remove = TRUE)
}
\arguments{
\item{expr}{Expression to evaluate.}

\item{tmpdir}{Temporary directory to evaluate it in. By default
a new temporary direcotory is created.}

\item{remove}{Whether to remove the temporary directory
after the expression is evaluated.}
}
\value{
Return value of the expression.
}
\description{
Run an expression in a temporary directory
}
\keyword{internal}
