% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_wd}
\alias{with_wd}
\title{Run an expression with a different working directory}
\usage{
with_wd(wd, expr)
}
\arguments{
\item{wd}{Temporary working directory to use.}

\item{expr}{Expression to evaluate.}
}
\value{
Return value of the expression.
}
\description{
Run an expression with a different working directory
}
\keyword{internal}

