% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_package}
\alias{add_package}
\title{Add or update a package in CRAN @ GitHub}
\usage{
add_package(package)
}
\arguments{
\item{package}{Name of the package}
}
\value{
Invisible `TRUE` if the package was successfully updated,
  `FALSE` otherwise.
}
\description{
The update is performed based on the current state of the
crandb database (See \url{https://github.com/metacran/crandb}.)
}
\details{
All versions that are missing form the GitHub repository
at \url{https://github.com/cran}, are added.

If the package is missing from the GitHub mirror completely,
then it is created.
}

