% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb.R
\name{get_all_cran_packages}
\alias{get_all_cran_packages}
\title{All active packages on CRAN}
\usage{
get_all_cran_packages()
}
\value{
Character vector of all active CRAN packages.
}
\description{
All active packages on CRAN
}
\keyword{internal}
