% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_gh_token}
\alias{get_gh_token}
\title{Get the token to be used for GitHub API calls}
\usage{
get_gh_token()
}
\value{
Character scalar, the token, or \code{NA} is the
  \code{GITHUB_TOKEN} environment variable is not set.
}
\description{
It is taken from the \code{GITHUB_TOKEN} environment
variable.
}
\keyword{internal}
