% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\title{Run a git command using system git}
\usage{
git(..., env = character())
}
\arguments{
\item{...}{Arguments to pass to git.}

\item{env}{Named character vector, environment variables
to set for git.}
}
\value{
A list containing the \code{status}, \code{stdout},
  \code{stderr} of the system git command.
}
\description{
System git must be properly installed and it must be in the
\code{PATH}.
}
\keyword{internal}
